Move file handling functions into a file utils
