+++ spidermon/commands/setup.py
@@ -3,7 +3,8 @@ import click
 import spidermon
 from scrapy.utils.project import inside_project
 from spidermon.commands.prompts import monitor_prompts
-from spidermon.utils.commands import build_monitors_strings, create_file, find_monitors, render_monitors
+from spidermon.utils.commands import build_monitors_strings, find_monitors
+from spidermon.utils.file import create_file, render_file
 
 @click.command('setup', help="Setup the monitors from the Scrapy Monitor Suite.")
 def setup():
@@ -28,6 +29,10 @@ def setup():
 
     # create and build file with monitors
     filename = create_file('monitor_suite.py.tmpl')
-    render_monitors(filename, monitors_list, monitors_imports)
+    render_file(
+        filename,
+        monitors_list=monitors_list,
+        monitors_imports=monitors_imports
+    )
 
     click.echo(monitor_prompts['response'])
